Avoid using std.concurrency for runTaskDistH.
std.concurrency is highly inefficient and fully relies on the GC to clean up resources (including semaphore OS handles on Windows), resulting in considerable time and memory overhead. Using a channel to make this more streamlined.

This also allows removing the task wrapping that used to be necessary when the function is called outside of a task.