Fix TCPConnection double-release after read timeouts.
The existing releaseHandle() method only reset the m_socket field if the release was the final one. This meant that calling close() during a pending read would leave the field set and thus cause another release from within the destructor. If that release came before the read operation was finished, this resulted in the underlying eventcode socket slot to be cleared before all operations had actually finished, consequently causing a crash in eventcore.