Work around unfair event object behavior in eventcore.
The current trigger behavior for "single" notifications sometimes resulted in all tasks ending up in the same worker thread, even if that thread was constantly busy. By notifying all threads, we do introduce a little overhead for each task, but guarantee that a worker thread that is not currently busy takes the task.