Fix SyslogLogger to be generally thread and task safe. See #176.
Uses an asynchronous writer thread to write the log output and avoids any yielding within the thread/task that performs the logging call.

Note that this change breaks the API to create a SyslogLogger instance, but since the previous implementation was so fundamentally broken, we can assume that nobody was using it.