Fix bogus FileStream destruction warning.
The warning in its previous form was too general. It is only an actual problem if the last reference to a FileStream is leaked to the GC, as long as it is ensured that the file gets closed before the GC attempts to finalize the last copy, everything is safe.

See #412